itc_poc_server/requirements.txt
alembic==1.1.0
aniso8601==7.0.0
Click==7.0
f==0.0.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.3
marshmallow-sqlalchemy==0.19.0
mysqlclient==1.4.4
pkg-resources==0.0.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
XlsxWriter==1.2.0
